+++ CHANGELOG.md
@@ -13,7 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Support for escape and unicode characters [#98](https://github.com/ufirstgroup/ymlr/pull/98)
-- Refactor `Ymlr.Encode` to make it faster.
+- Refactor `Ymlr.Encode` to make it faster [#171](https://github.com/ufirstgroup/ymlr/pull/171).
+- Handle special floats `.Inf` and `.Nan`. [#170](https://github.com/ufirstgroup/ymlr/issues/170), [#172](https://github.com/ufirstgroup/ymlr/issues/172)
 
 <!--------------------- Don't add new entries after this line --------------------->