+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 <!-- ### Added | Changed | Deprecated | Removed | Fixed | Security -->
 
+<!--------------------- Don't add new entries after this line --------------------->
+
+## [5.1.0] - 2023-12-27
+
 ### Added
 
 - Support for escape and unicode characters [#98](https://github.com/ufirstgroup/ymlr/pull/98)
@@ -20,8 +24,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Put single quotes around indicator chars when it is the entire value. [#177](https://github.com/ufirstgroup/ymlr/issues/177), [#178](https://github.com/ufirstgroup/ymlr/pull/178)
-
-<!--------------------- Don't add new entries after this line --------------------->
 
 ## [5.0.0] - 2023-11-03